Rename Helpers.js to helpers.js - Format useIpenWeatherApi file
