+++ src/hooks/useOpenWeatherApi.js
@@ -1,6 +1,6 @@
 import { useState, useEffect } from 'react';
 import useFetch from './useFetch';
-import { getIconClass, makeDate } from '../Helpers';
+import { getIconClass, makeDate } from '../helpers';
 
 const API_KEY = import.meta.env.VITE_API_KEY;
 
@@ -13,29 +13,26 @@ function useWeather(query) {
   const [weather, setWeather] = useState(data);
   const [temps, setTemps] = useState(data);
 
-  useEffect(
-    () => {
-      if (data !== null) {
-        setWeather({
-          city: data.name,
-          country: data.sys.country,
-          date: makeDate(),
-          weather: data.weather[0].main,
-          description: data.weather[0].description,
-          icon: getIconClass(data.weather[0].icon),
-          pressure: data.main.pressure,
-          humidity: data.main.humidity,
-        });
-        setTemps({
-          temp: data.main.temp,
-          max: data.main['temp_max'],
-          min: data.main['temp_min'],
-          feelsLike: data.main['feels_like'],
-        });
-      }
-    },
-    [data]
-  );
+  useEffect(() => {
+    if (data !== null) {
+      setWeather({
+        city: data.name,
+        country: data.sys.country,
+        date: makeDate(),
+        weather: data.weather[0].main,
+        description: data.weather[0].description,
+        icon: getIconClass(data.weather[0].icon),
+        pressure: data.main.pressure,
+        humidity: data.main.humidity,
+      });
+      setTemps({
+        temp: data.main.temp,
+        max: data.main['temp_max'],
+        min: data.main['temp_min'],
+        feelsLike: data.main['feels_like'],
+      });
+    }
+  }, [data]);
   return { weather, temps, isLoading, hasError, errorMessage };
 }
